requirements.txt
streamlit
streamlit-cropperjs
pandas
numpy
plotly
altair
faiss-cpu
umap-learn
scikit-learn
imageio
pillow
opencv-python
transformers
torch
torchvision
sentencepiece
